backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
Pillow
djoser==2.1.0
requests==2.30.0
django-filter==23.2
drf-base64==2.0
python-dotenv
asgiref==3.3.2
gunicorn==20.0.4
psycopg2-binary==2.9.6
sqlparse==0.3.1
